move sparse_center_data to linear_model.base
