+++ sklearn/linear_model/base.py
@@ -23,6 +23,9 @@ from ..base import RegressorMixin
 from ..utils.extmath import safe_sparse_dot
 from ..utils import array2d, as_float_array, safe_asarray
 from ..utils.fixes import lsqr
+from ..utils.sparsefuncs import csc_mean_variance_axis0, \
+                                inplace_csc_column_scale
+from cd_fast import sparse_std
 
 
 ###
@@ -32,6 +35,36 @@ from ..utils.fixes import lsqr
 
 ### TODO: bayesian_ridge_regression and bayesian_regression_ard
 ### should be squashed into its respective objects.
+
+
+def sparse_center_data(X, y, fit_intercept, normalize=False):
+    """
+    Compute informations needed to center data to have mean zero along
+    axis 0. Be aware that X will not be centered since it would break
+    the sparsity, but will be normalized if asked so.
+    """
+    X_data = np.array(X.data, np.float64)
+    if fit_intercept:
+        # copy if 'normalize' is True or X is not a csc matrix
+        X = sp.csc_matrix(X, copy=normalize)
+        X_mean, X_std = csc_mean_variance_axis0(X)
+        if normalize:
+            X_std = sparse_std(
+                X.shape[0], X.shape[1],
+                X_data, X.indices, X.indptr, X_mean)
+            X_std[X_std == 0] = 1
+            inplace_csc_column_scale(X, X_std)
+        else:
+            X_std = np.ones(X.shape[1])
+        y_mean = y.mean(axis=0)
+        y = y - y_mean
+    else:
+        X_mean = np.zeros(X.shape[1])
+        X_std = np.ones(X.shape[1])
+        y_mean = 0. if y.ndim == 1 else np.zeros(y.shape[1], dtype=X.dtype)
+
+    X_data = np.array(X.data, np.float64)
+    return X_data, y, X_mean, y_mean, X_std
 
 
 def center_data(X, y, fit_intercept, normalize=False, copy=True):
